git-utils: isolate async status guard
